Reorganizing the structure of the project01
