+++ classes/exception_classes.py
@@ -1,0 +1,8 @@
+class LenNumberError(Exception):
+    def __init__(self, txt):
+        self.txt = txt
+
+
+class NameError(Exception):
+    def __init__(self, txt):
+        self.txt = txt
