+++ functions/file_operations.py
@@ -1,0 +1,28 @@
+from csv import DictReader, DictWriter
+
+def create_file(file_name):
+    # with - Менеджер контекста
+    with open(file_name, "w", encoding='utf-8') as data:
+        f_writer = DictWriter(data, fieldnames=['Имя', 'Фамилия', 'Телефон'])
+        f_writer.writeheader()
+
+
+def read_file(file_name):
+    with open(file_name, "r", encoding='utf-8') as data:
+        f_reader = DictReader(data)
+        return list(f_reader)
+
+
+def write_file(file_name, lst):
+    res = read_file(file_name)
+    for el in res:
+        if el["Телефон"] == str(lst[2]):
+            print("Такой телофон уже есть")
+            return
+
+    obj = {"Имя": lst[0], "Фамилия": lst[1], "Телефон": lst[2]}
+    res.append(obj)
+    with open(file_name, "w", encoding='utf-8', newline='') as data:
+        f_writer = DictWriter(data, fieldnames=['Имя', 'Фамилия', 'Телефон'])
+        f_writer.writeheader()
+        f_writer.writerows(res)
