+++ functions/get_info.py
@@ -1,0 +1,31 @@
+def get_info():
+    is_valid_first_name = False
+    while not is_valid_first_name:
+        try:
+            first_name = input("Введите имя: ")
+            if len(first_name) < 2:
+                raise NameError("Не валидное имя")
+            else:
+                is_valid_first_name = True
+        except NameError as err:
+            print(err)
+            continue
+
+    last_name = "Иванов"
+
+    is_valid_phone = False
+    while not is_valid_phone:
+        try:
+            phone_number = int(input("Введите номер: "))
+            if len(str(phone_number)) != 11:
+                raise LenNumberError("Неверная длина номера")
+            else:
+                is_valid_phone = True
+        except ValueError:
+            print("Не валидный номер!!!")
+            continue
+        except LenNumberError as err:
+            print(err)
+            continue
+
+    return [first_name, last_name, phone_number]
