+++ functions/main.py
@@ -1,0 +1,18 @@
+from os.path import exists
+from functions.get_info import get_info
+from functions.file_operations import create_file, read_file, write_file
+
+def main(file_name):
+    while True:
+        command = input("Введите команду: ")
+        if command == 'q':
+            break
+        elif command == 'w':
+            if not exists(file_name):
+                create_file(file_name)
+            write_file(file_name, get_info())
+        elif command == 'r':
+            if not exists(file_name):
+                print("Файл отсутствует. Создайте его")
+                continue
+            print(*read_file(file_name))
