+++ homework.py
@@ -1,0 +1,20 @@
+# Создать телефонный справочник с возможностью импорта и экспорта данных в
+# формате .txt. Фамилия, имя, отчество, номер телефона - данные, которые 
+# должны находиться в файле.
+# 1. Программа должна выводить данные
+# 2. Программа должна сохранять данные в текстовом файле
+# 3. Пользователь может ввести одну из характеристик для поиска определенной
+# записи(Например имя или фамилию человека)
+# 4. Использование функций. Ваша программа не должна быть линейной
+
+from classes.exception_classes import LenNumberError, NameError
+from functions.main import main
+
+file_name = 'phone.csv'
+
+main()
+
+# Дополнить справочник возможностью
+# копирования данных из одного файла в другой.
+# Пользователь вводит номер строки, которую
+# необходимо перенести из одного файла в другой.
